backend/requirements.txt
Django==3.2
djangorestframework==3.12.0
django-filter==2.4.0
django-rest-knox==3.1.4
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
drf-psq==1.1.0
coreapi==2.3.3
markdown==3.0.1 
pygments==2.7.4
Pillow==8.2.0
sorl-thumbnail==12.7.0
django-cors-headers==3.4.0
ptvsd==4.3.2
psycopg2==2.8.6
djangorestframework-recaptcha==0.2.0
celery==5.0.0
